+++ R/frank.R
@@ -32,11 +32,14 @@ frankv = function(x, cols=seq_along(x), order=1L, na.last=TRUE, ties.method=c("a
     nas  = x[[ncol(x)]]
   }
   if (ties.method == "random") {
-    v = stats::runif(nrow(x))
     if (is.na(na.last)) {
       idx = which_(nas, FALSE)
-      set(x, idx, '..stats_runif..', v[idx])
-    } else set(x, NULL, '..stats_runif..', v)
+      n = length(idx)  # changed in #4243 to match base R to pass test 1369
+    } else {
+      idx = NULL
+      n = nrow(x)
+    }
+    set(x, idx, '..stats_runif..', stats::runif(n))
     order = if (length(order) == 1L) c(rep(order, length(cols)), 1L) else c(order, 1L)
     cols = c(cols, ncol(x))
   }